Fixes: #306 (Set up Status Open Bounties for Giveth)
I fixed a small readme in the typo in order to be able to make a pull request. A pull request is needed for *bounties* to pay out. Since the issue is dealing with a discussion only I chose to commit to Readme as a safe choice for this bounty testrun