[WIP] Save Unfinished Milestones, Campaigns and DACs
This is work in progress to address issue #522. A new file was created, Draft.jsx, which
defines a new button that allows to save drafts, as well as the draft saving functionality
itself. It was created with the objective of being reusable in different views. It is not a
component itself, but rather uses the context of the component that imports it. So far, the
saving functionality has been mostly implemented, as well as the event listeners that enable
and disable the button, and the button itself. The loading functionality is still missing,
and also there's a small bug that enables the button when it should be disabled.

There's a shortcoming to this approach that needs to be tested: LocalStorage only allows
for 5MB of storage per domain. In some browsers, this can be considerably less. This means
that large images will not be able to be saved, especially when having multiple drafts. A
failsafe has to be implemented for these cases.