Draft loading (#522)
Implemented loading functionality. Since the form is an uncontrolled component (see https://reactjs.org/docs/uncontrolled-components.html), I had to access the DOM directly using references. This turned out to be very difficult because the form is loaded *after* componentDidMount. So after a lot of research and trial and error, I ended up calling the loading function in the onValid event of the form. Unfortunately, this event is initially called 11 times and there doesn't seem to be any possible way to prevent this, since the state hasn't been initialized yet.