Fix issue#645,Delegating to milestones or campaign
The problem:
Right now there are just too many options in this drop down and its confusing cause you dont know what campaigns each milestone listed is in... you only get the title... so you might think you are delegating to a milestone in the giveth communications campaign, but actually you are delegating to the Bloom Campaign (this happened to me!)

proposal:

Make the drop down have 2 levels. The first level being the Campaign and then when you select a campaign, you get to select the milestones in that selected campaign PLUS an option to delegate directly to the campaign itself.